bb-helper: file_stream: Add integration tests
Make unit tests integration tests.
